+++ libft/src/ft_mem/ft_memcmp.c
@@ -12,9 +12,9 @@
 
 #include"libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, size_t n)
+int	ft_memcmp(const void *s1, const void *s2, int n)
 {
-	size_t				i;
+	int				i;
 	const unsigned char	*str1;
 	const unsigned char	*str2;
