+++ libft/src/ft_mem/ft_memset.c
@@ -12,9 +12,9 @@
 
 #include"libft.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, int len)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < len)
